Added Books controller
Moved all books logic into books controller
